app/requirements.txt
wheel
gunicorn
flask
flask-mail
boto3
pyyaml
requests
atomicwrites
